backend/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.3
Django==3.2.16
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
psycopg2-binary==2.9.5
pycparser==2.21
pytz==2022.5
sqlparse==0.4.3
typing_extensions==4.4.0
whitenoise==6.2.0
django-admin-notification==0.2
django-storages==1.13.2
boto3==1.26.77
python-decouple==3.8